# Swagger.yaml
...
openapi: "3.0.0"
info:
  title: "No Appearance API"
  description: "Esta Api * Creacion de Cuenta * Eleccion de preferencias * Menu Principal\n\
    \n* Lista de historial de chat * Mensajeria * Lista de usuarios reportados * Reporte\
    \ del usuario * Reportar usuario\n\n* Agregar Usuario * Editar datos de usuario\
    \ * Editar perfil * Editar categorias del usuario * Ver información del usuario\n\
    * Agregar categorias * Eliminar categorias * Editar categorias * Agregar subcategorias\
    \ * Eliminar subcategorias * Editar subcategoria"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
  version: "1.0.0-oas3"
...
servers:
  - url: "https://virtserver.swaggerhub.com/Ro-Nel/Pru/1.0.0-oas3"
    description: "SwaggerHub API Auto Mocking"
  - url: "https://virtserver.swaggerhub.com/bipersonal/NoAppearence/1.0.0"
    description: "SwaggerHub API Auto Mocking"
  - url: "https://virtserver.swaggerhub.com/bipersonal/CATEGORY/1.0.0"
    description: "SwaggerHub API Auto Mocking"
...
paths:
  /users:
    post:
      tags:
        - "users"
      summary: "Add a new user to the system"
      operationId: "addUser"
      requestBody:
        $ref: "#/components/requestBodies/Users"
      responses:
        "405":
          description: "Invalid input"
...
components:
  schemas:
...
    Users:
      type: "object"
      properties:
        userId:
          type: "string"
        userType:
          type: "string"
        name:
          type: "string"
        surname:
          type: "string"
        birthdate:
          type: "string"
        gender:
          type: "boolean"
        email:
          type: "string"
        password:
          type: "string"
        userPhoto:
          type: "string"
      xml:
        name: "Users"
  requestBodies:
...
    Users:
      description: "User object that needs to be added to the system"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Users"
        application/xml:
          schema:
            $ref: "#/components/schemas/Users"
      required: true
...